Modelling/Deploy/search-bar/requirements.txt
Flask
joblib
dill
numpy
pandas
scikit-learn
google-cloud-bigquery==3.11.0
db-dtypes
gunicorn==20.1.0